+++ libs/admin/src/lib/admin-panel/components/organization/forms/organization-form/organization-form.component.html
@@ -34,9 +34,9 @@
       <h2>Company Logo</h2>
       <h5>Upload company Logo</h5>
 
-      <drop-cropper ratio="square" [storagePath]="storagePath" [form]="form.get('logo')">
+      <image-uploader ratio="square" [storagePath]="storagePath" [form]="form.get('logo')">
         <h3 title>Logo</h3>
-      </drop-cropper>
+      </image-uploader>
     </article>
   </section>
