+++ libs/admin/src/lib/admin-panel/components/organization/forms/organization-form/organization-form.module.ts
@@ -17,7 +17,7 @@ import { MatInputModule } from '@angular/material/input';
 // Modules
 import { OrganizationFormAddressModule } from '@blockframes/organization/forms/organization-form-address/organization-form-address.module';
 import { OrganizationFormDescriptionModule } from '@blockframes/organization/forms/organization-form-description/organization-form-description.module';
-import { CropperModule } from '@blockframes/media/image/cropper/cropper.module';
+import { ImageUploaderModule } from '@blockframes/media/image/uploader/uploader.module';
 
 
 @NgModule({
@@ -37,7 +37,7 @@ import { CropperModule } from '@blockframes/media/image/cropper/cropper.module';
     // Blockframes
     OrganizationFormAddressModule,
     OrganizationFormDescriptionModule,
-    CropperModule
+    ImageUploaderModule
   ],
   declarations: [AdminOrganizationFormComponent],
   exports: [AdminOrganizationFormComponent]
