+++ libs/media/src/lib/file/explorer/cropper-dialog/cropper-dialog.component.html
@@ -1,9 +1,9 @@
 <h2>Edit Image</h2>
 <mat-divider></mat-divider>
 <form [formGroup]="data.form" fxLayout="column" fxLayoutGap="24px">
-  <drop-cropper [ratio]="data.ratio" [storagePath]="data.storagePath" [form]="data.form">
+  <image-uploader [ratio]="data.ratio" [storagePath]="data.storagePath" [form]="data.form">
     <h3 title>Image</h3>
-  </drop-cropper>
+  </image-uploader>
   <section fxLayout="column" fxLayoutAlign="center center">
     <button mat-stroked-button color="primary" (click)="upload()" [disabled]="data.form.invalid">
       <span>Update</span>
