+++ libs/media/src/lib/file/explorer/explorer.model.ts
@@ -1,6 +1,6 @@
 import { Privacy } from "@blockframes/utils/file-sanitizer";
 import { FormList } from "@blockframes/utils/form";
-import { MediaRatioType } from '../../image/cropper/cropper.component';
+import { MediaRatioType } from '../../image/uploader/uploader.component';
 import { HostedMediaWithMetadataForm } from '@blockframes/media/form/media-with-metadata.form';
 import { HostedMediaForm } from '@blockframes/media/form/media.form';
 import { isMediaForm } from "@blockframes/media/+state/media.model";
