+++ libs/media/src/lib/file/explorer/single-file-view/single-file-view.component.html
@@ -1,9 +1,9 @@
 <ng-container *ngIf="!!this.activeForm">
   <section fxLayout="column" fxLayoutGap="24px">
     <ng-container *ngIf="activeDirectory.type === 'image'">
-      <drop-cropper [ratio]="activeDirectory.ratio" [storagePath]="activeDirectory.storagePath" [form]="getSingleMediaForm()">
+      <image-uploader [ratio]="activeDirectory.ratio" [storagePath]="activeDirectory.storagePath" [form]="getSingleMediaForm()">
         <h3 title>{{ activeDirectory.name }}</h3>
-      </drop-cropper>
+      </image-uploader>
     </ng-container>
 
     <ng-container *ngIf="activeDirectory.type === 'file'">
