+++ libs/media/src/lib/image/uploader/uploader.component.scss
