+++ libs/organization/src/lib/organization/forms/organization-form/organization-form.component.html
@@ -38,9 +38,9 @@
         <mat-card-subtitle class="mat-caption">Upload your Logo</mat-card-subtitle>
       </mat-card-header>
       <mat-card-content fxLayout="column">
-        <drop-cropper ratio="square" [storagePath]="orgId | ref: 'organization' : 'logo'" [form]="form.logo">
+        <image-uploader ratio="square" [storagePath]="orgId | ref: 'organization' : 'logo'" [form]="form.logo">
           <h3 title>Logo</h3>
-        </drop-cropper>
+        </image-uploader>
       </mat-card-content>
     </mat-card>
     <!-- Legal Entity -->
