+++ libs/organization/src/lib/organization/no-organization.module.ts
@@ -5,7 +5,7 @@ import { ReactiveFormsModule, FormsModule } from '@angular/forms';
 import { FlexLayoutModule } from '@angular/flex-layout';
 
 // Modules
-import { CropperModule } from '@blockframes/media/image/cropper/cropper.module';
+import { ImageUploaderModule } from '@blockframes/media/image/uploader/uploader.module';
 import { OrganizationFormModule } from './forms/organization-form/organization-form.module';
 import { ImageModule } from '@blockframes/media/image/directives/image.module';
 import { AlgoliaAutocompleteModule } from '@blockframes/ui/algolia/autocomplete/algolia-autocomplete.module';
@@ -83,7 +83,7 @@ export const noOrganizationRoutes: Routes = [
     ReactiveFormsModule,
     FormsModule,
     FlexLayoutModule,
-    CropperModule,
+    ImageUploaderModule,
     OrganizationFormModule,
     ImageModule,
     AlgoliaAutocompleteModule,
