+++ libs/organization/src/lib/organization/organization.module.ts
@@ -7,7 +7,7 @@ import { FlexLayoutModule } from '@angular/flex-layout';
 // Library
 import { FileUploaderModule } from '@blockframes/media/file/file-uploader/file-uploader.module';
 import { PasswordConfirmModule } from '@blockframes/ui/form/password-confirm/password-confirm.module';
-import { CropperModule } from '@blockframes/media/image/cropper/cropper.module';
+import { ImageUploaderModule } from '@blockframes/media/image/uploader/uploader.module';
 import { ImageModule } from '@blockframes/media/image/directives/image.module';
 import { OrganizationFormModule } from './forms/organization-form/organization-form.module';
 import { EditableSidenavModule } from '@blockframes/ui/editable-sidenav/editable-sidenav.module';
@@ -70,7 +70,7 @@ export const organizationRoutes: Routes = [
     AvatarListModule,
     FileUploaderModule,
     PasswordConfirmModule,
-    CropperModule,
+    ImageUploaderModule,
     ImageModule,
     OrganizationFormModule,
     TableFilterModule,
