+++ libs/user/src/lib/auth/forms/profile/profile.component.html
@@ -37,8 +37,8 @@
     <mat-card-subtitle>Choose your avatar</mat-card-subtitle>
   </mat-card-header>
   <mat-card-content>
-    <drop-cropper ratio="square" [storagePath]="uid | ref: 'profile' : 'avatar'" [form]="form.avatar">
+    <image-uploader ratio="square" [storagePath]="uid | ref: 'profile' : 'avatar'" [form]="form.avatar">
       <h3 title>Profile Image</h3>
-    </drop-cropper>
+    </image-uploader>
   </mat-card-content>
 </mat-card>
